backend/requirements.txt
django==4.2.7
djangorestframework==3.15.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
stripe==7.6.0
python-decouple==3.8
gunicorn==23.0.0
